Moved prepare_work_graph_copy and connected functions to separate file
